ecourseapisv1/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cloudinary==1.44.1
Django==5.2.7
django-ckeditor==6.7.3
django-js-asset==3.1.2
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pillow==12.0.0
PyMySQL==1.1.2
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
